Move `dataframe` and `dataframe_functon` into `core_integration` teset binary
