+++ datafusion/core/tests/dataframe/mod.rs
@@ -15,15 +15,5 @@
 // specific language governing permissions and limitations
 // under the License.
 
-/// Run all tests that are found in the `sql` directory
-mod sql;
-
-/// Run all tests that are found in the `dataframe` directory
 mod dataframe;
-
-#[cfg(test)]
-#[ctor::ctor]
-fn init() {
-    // Enable RUST_LOG logging configuration for test
-    let _ = env_logger::try_init();
-}
+mod dataframe_functions;
